% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{customize_style}
\alias{customize_style}
\title{Add custom style to markdown text}
\usage{
customize_style(text, style = "")
}
\arguments{
\item{text}{The text to be adjusted}

\item{style}{The name of the custom style}
}
\value{
The pandoc-flavored markdown string
}
\description{
Add custom style to markdown text
}
