% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_tmpl_name}
\alias{get_tmpl_name}
\title{Get template file name based on current configuration}
\usage{
get_tmpl_name(part, customer_id, report_type = NULL, ext = "Rmd")
}
\arguments{
\item{part}{The report part}

\item{customer_id}{The customer identifier}

\item{report_type}{The report type, and default to \code{\link{NULL}}}

\item{ext}{The file extension of the template, the default value is
\code{Rmd}, namely R markdown file}
}
\description{
Get template file name based on current configuration
}
