% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_heading}
\alias{render_heading}
\title{Render heading markdown according to level}
\usage{
render_heading(heading, hlevel = 1)
}
\arguments{
\item{heading}{Heading text}

\item{hlevel}{Heading level, default to 1}
}
\value{
The corresponding heading level markdown content
}
\description{
Render heading markdown according to level
}
