% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_heading_content}
\alias{render_heading_content}
\title{Render heading and content section}
\usage{
render_heading_content(heading, content, hlevel = 1, style = "",
  glue = FALSE, ...)
}
\arguments{
\item{heading}{The section title}

\item{content}{The section content}

\item{hlevel}{The heading level, default to 1}

\item{style}{The name of the custom style}

\item{glue}{Will the content be evaluated by \code{\link[stringr]{str_glue}}?}

\item{...}{Additional arguments passed to \code{\link[stringr]{str_glue}}}
}
\value{
The markdown string to render as a section
}
\description{
Render heading and content section
}
