% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_part_body}
\alias{render_part_body}
\title{Render markdown for body (special) part}
\usage{
render_part_body(heading, content, ab_ids)
}
\arguments{
\item{heading}{The heading text of the body part}

\item{content}{The raw R markdown file loaded from template}

\item{ab_ids}{Full character vector containing all the ability identifiers to
be reported}
}
\description{
Render markdown for body (special) part
}
