% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_part_normal}
\alias{render_part_normal}
\title{Render markdown for normal part}
\usage{
render_part_normal(text)
}
\arguments{
\item{text}{The raw R markdown file loaded from template}
}
\description{
Render markdown for normal part
}
